% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataimport.R
\name{combine_files}
\alias{combine_files}
\title{Combine files from same rat into a tibble}
\usage{
combine_files(filepaths, baselinedays, exposuredays, zeitgebertime = 7)
}
\arguments{
\item{filepaths}{a list of files}

\item{baselinedays}{number of days in baseline condition, defaults to 4}

\item{exposuredays}{number of days in exposure condition, defaults to 5}

\item{zeitgebertime}{number indicating start of light period, defaults to 7 (= 07:00)}
}
\description{
Takes a list of .txt files from the same rat and produces a tibble with all variables
}
