% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataimport.R
\name{create_data}
\alias{create_data}
\title{Load data set}
\usage{
create_data(filepaths, baseline, exposure = 7, baselinestart,
  maxbaseline = 5)
}
\arguments{
\item{filepaths}{A named list of character vectors, indicating the paths to all
the files of the same animal.}

\item{baseline}{A named list of the number of baselinedays for each animal,
or a number if it is the same for each animal, no default}

\item{exposure}{A number, indicating length of exposure period in days, defaults to 5}

\item{baselinestart}{A named list indicating start of baseline in hours (e.g. 7 = 07:00) for each animal,
or a number if it is the same for each animal}

\item{maxbaseline}{A number of maximum baseline days in experiment, defaults to 5}
}
\value{
a tibble
}
\description{
Load list of .txt files from data quest art. All the files must be named VARIABLE_GROUP_ID,
where ID should be numbers only. This makes sure the data set is loaded, the new variable
is properly named, and the data is aligned to the correct ID. The filepaths argument must be a list, where every element of the list contains all files for
one animal. The other arguements may be a single number (if same value for each animal), or a list
of the same length as filepaths, with one value for each animal.
}
\examples{
animals <- list(B8 = c("./Data/Activity_B_8.txt", "./Data/Bodytemp_B_8.txt"),
                W12 = c("./Data/Activity_W_12.txt", "./Data/Bodytemp_W_12.txt"))
baselinedays <- list(B8 = 5, W12 = 4)

create_data(animals,
            baseline = baselinedays,
            exposure = 7,
            baselinestart = 7,
            maxbaseline = 5)

}
