% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataformat.R
\name{format_normalize}
\alias{format_normalize}
\title{Normalizes a variable to specified experimental period}
\usage{
format_normalize(.data, ID_var, variable, phase_var = Experiment_period,
  phase_value = Baseline, add = 0)
}
\arguments{
\item{.data}{a tibble}

\item{ID_var}{the ID variable, for grouping}

\item{variable}{name of the variable that is going to be transformed}

\item{phase_var}{name of the variable containing experimental phases, defaults to Experiment_period}

\item{phase_value}{name of the value of phase_var which you normalizes towards, defaults to Baseline}

\item{add}{Optional argument, a number to add to every value if you want the mean to be something
other than 0. Defaults to 0.}
}
\value{
a tibble, where one variable is transformed
}
\description{
Subtracting mean of the specified period from every value in the variable that's being transformed, i.e.
mean of the baseline period
}
\examples{
Normalizing variable bodytemperature to the Baseline period,
contained in variable Experiment_period in data set d

d \%>\%
  format_normalize(variable = bodytemperature,
                   ID_var = Id,
                   phase_var = Experiment_period,
                   phase_value = Baseline,
                   add = 0)

}
