% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataimport.R
\name{load_single_file}
\alias{load_single_file}
\title{Load a single .txt data file from "data quest art" software}
\usage{
load_single_file(path, baselinedays = 4, exposuredays = 5,
  zeitgebertime = 7)
}
\arguments{
\item{path}{path to file}

\item{baselinedays}{number of days in baseline condition, defaults to 4}

\item{exposuredays}{number of days in exposure condition, defaults to 5}

\item{zeitgebertime}{number indicating start of light period, defaults to 7 (= 07:00)}
}
\description{
Load a single .txt file with two columns, and creates column names.
File name should be "Variablename_groupname_ratID.txt", as these are used
to create column names, as well as group and id columns.
}
