% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataimport.R
\name{create_data}
\alias{create_data}
\title{Load data set}
\usage{
create_data(filepaths, baseline, exposure = 7, zeitgebertime,
  maxbaseline = 5)
}
\arguments{
\item{filepaths}{A named list of character vectors, indicating the paths to all
the files of the same animal.}

\item{baseline}{A named list of the number of baselinedays for each animal,
or a number if it is the same for each animal, no default}

\item{exposure}{A number, indicating length of exposure period in days, defaults to 5}

\item{zeitgebertime}{A named list indicating start of light period for each animal,
or a number if it is the same for each animal}

\item{maxbaseline}{A number of maximum baseline days in experiment, defaults to 5}
}
\value{
a tibble
}
\description{
Load data set
}
\examples{
animals <- list(B8 = c("./Data/Activity_B_8.txt", "./Data/Bodytemp_B_8.txt"),
                W12 = c("./Data/Activity_W_12.txt", "./Data/Bodytemp_W_12.txt"))
baselinedays <- list(B8 = 5, W12 = 4)

create_data(animals,
            baseline = baselinedays,
            exposure = 7,
            zeitgebertime = 7,
            maxbaseline = 5)

}
